+++ crates/astria-sequencer/src/app/tests_app/mempool.rs
@@ -728,8 +728,9 @@ async fn proposer_flow_included_transactions_sent_to_mempool() {
         2,
         "removal cache should have 2 txs"
     );
-    let TransactionStatus::Removed(RemovalReason::IncludedInBlock(height_1)) =
-        fixture.mempool().transaction_status(tx_1_id).await.unwrap()
+    let TransactionStatus::Removed(RemovalReason::IncludedInBlock {
+        height: height_1, ..
+    }) = fixture.mempool().transaction_status(tx_1_id).await.unwrap()
     else {
         panic!("tx_1 should be removed");
     };
@@ -738,8 +739,9 @@ async fn proposer_flow_included_transactions_sent_to_mempool() {
         height.value(),
         "tx_1 should be removed from mempool with height {height}"
     );
-    let TransactionStatus::Removed(RemovalReason::IncludedInBlock(height_2)) =
-        fixture.mempool().transaction_status(tx_2_id).await.unwrap()
+    let TransactionStatus::Removed(RemovalReason::IncludedInBlock {
+        height: height_2, ..
+    }) = fixture.mempool().transaction_status(tx_2_id).await.unwrap()
     else {
         panic!("tx_2 should be removed");
     };
@@ -883,8 +885,9 @@ async fn non_proposer_validator_flow_included_transactions_sent_to_mempool() {
         2,
         "removal cache should have 2 txs"
     );
-    let TransactionStatus::Removed(RemovalReason::IncludedInBlock(height_1)) =
-        fixture.mempool().transaction_status(tx_1_id).await.unwrap()
+    let TransactionStatus::Removed(RemovalReason::IncludedInBlock {
+        height: height_1, ..
+    }) = fixture.mempool().transaction_status(tx_1_id).await.unwrap()
     else {
         panic!("tx_1 should be removed");
     };
@@ -893,8 +896,9 @@ async fn non_proposer_validator_flow_included_transactions_sent_to_mempool() {
         height.value(),
         "tx_1 should be removed from mempool with height {height}"
     );
-    let TransactionStatus::Removed(RemovalReason::IncludedInBlock(height_2)) =
-        fixture.mempool().transaction_status(tx_2_id).await.unwrap()
+    let TransactionStatus::Removed(RemovalReason::IncludedInBlock {
+        height: height_2, ..
+    }) = fixture.mempool().transaction_status(tx_2_id).await.unwrap()
     else {
         panic!("tx_2 should be removed");
     };
@@ -905,6 +909,7 @@ async fn non_proposer_validator_flow_included_transactions_sent_to_mempool() {
     );
 }
 
+#[expect(clippy::too_many_lines, reason = "it's a test")]
 #[tokio::test]
 async fn non_validator_flow_included_transactions_sent_to_mempool() {
     // The flow of this test simulates a full node (running a mempool) that is not a validator
@@ -1007,8 +1012,9 @@ async fn non_validator_flow_included_transactions_sent_to_mempool() {
         2,
         "removal cache should have 2 txs"
     );
-    let TransactionStatus::Removed(RemovalReason::IncludedInBlock(height_1)) =
-        fixture.mempool().transaction_status(tx_1_id).await.unwrap()
+    let TransactionStatus::Removed(RemovalReason::IncludedInBlock {
+        height: height_1, ..
+    }) = fixture.mempool().transaction_status(tx_1_id).await.unwrap()
     else {
         panic!("tx_1 should be removed");
     };
@@ -1017,8 +1023,9 @@ async fn non_validator_flow_included_transactions_sent_to_mempool() {
         height.value(),
         "tx_1 should be removed from mempool with height {height}"
     );
-    let TransactionStatus::Removed(RemovalReason::IncludedInBlock(height_2)) =
-        fixture.mempool().transaction_status(tx_2_id).await.unwrap()
+    let TransactionStatus::Removed(RemovalReason::IncludedInBlock {
+        height: height_2, ..
+    }) = fixture.mempool().transaction_status(tx_2_id).await.unwrap()
     else {
         panic!("tx_2 should be removed");
     };